[refactor] Move main part of network package to anoher place
