+++ app/src/main/java/com/makentoshe/habrachan/common/network/converter/PostConverterFactory.kt
@@ -1,4 +1,4 @@
-package com.makentoshe.habrachan.common.model.network.postsalt
+package com.makentoshe.habrachan.common.network.converter
 
 import com.google.gson.Gson
 import com.makentoshe.habrachan.common.entity.post.PostResponse
