+++ app/src/main/java/com/makentoshe/habrachan/common/network/manager/HabrPostsManager.kt
@@ -1,7 +1,12 @@
-package com.makentoshe.habrachan.common.model.network.postsalt
+package com.makentoshe.habrachan.common.network.manager
 
 import com.makentoshe.habrachan.common.entity.post.PostResponse
 import com.makentoshe.habrachan.common.entity.posts.PostsResponse
+import com.makentoshe.habrachan.common.network.api.HabrPostsApi
+import com.makentoshe.habrachan.common.network.converter.PostConverterFactory
+import com.makentoshe.habrachan.common.network.converter.PostsConverterFactory
+import com.makentoshe.habrachan.common.network.request.GetPostRequest
+import com.makentoshe.habrachan.common.network.request.GetPostsRequest
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 import okhttp3.OkHttpClient
@@ -16,17 +21,22 @@ interface HabrPostsManager {
 
     fun getPost(request: GetPostRequest): Single<PostResponse>
 
-    companion object {
-        fun build(
-            client: OkHttpClient = OkHttpClient.Builder().build(),
-            baseUrl: String = "https://habr.com/"
-        ): HabrPostsManager {
-            val retrofit = Retrofit.Builder().client(client).baseUrl(baseUrl)
+    class Builder(private val client: OkHttpClient) {
+
+        private val baseUrl = "https://habr.com/"
+
+        private fun getRetrofit(): Retrofit {
+            return Retrofit.Builder().client(client).baseUrl(baseUrl)
                 .addCallAdapterFactory(RxJava2CallAdapterFactory.createWithScheduler(Schedulers.io()))
                 .addConverterFactory(PostsConverterFactory())
                 .addConverterFactory(PostConverterFactory())
                 .build()
+        }
+
+        fun build(): HabrPostsManager {
+            val retrofit = getRetrofit()
             return HabrPostsManagerImpl(retrofit.create(HabrPostsApi::class.java))
         }
     }
+
 }
